ai-services/api/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Dict, Any
import uvicorn
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="NEXUS Pro CRM - AI Services",
    description="AI-powered services for customer analytics and predictions",
    version="7.0.0"
)

# CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Models
class PredictionRequest(BaseModel):
    customer_id: str
    historical_data: List[Dict[str, Any]]

class ClassificationRequest(BaseModel):
    customer_data: Dict[str, Any]

class RecommendationRequest(BaseModel):
    customer_id: str
    context: Dict[str, Any]

class SentimentRequest(BaseModel):
    text: str

# Routes
@app.get("/")
async def root():
    return {
        "service": "NEXUS Pro CRM AI Services",
        "version": "7.0.0",
        "status": "operational"
    }

@app.get("/health")
async def health_check():
    return {
        "status": "healthy",
        "models_loaded": True,
        "services": ["prediction", "classification", "recommendation", "sentiment"]
    }

@app.post("/api/v1/predict")
async def predict_sales(request: PredictionRequest):
    """
    Predict future sales based on historical data
    """
    try:
        # TODO: Implement ML prediction model
        prediction = {
            "customer_id": request.customer_id,
            "predicted_value": 15000.00,
            "confidence": 0.87,
            "trend": "increasing",
            "factors": ["seasonality", "purchase_history", "engagement"]
        }
        return prediction
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/v1/classify")
async def classify_customer(request: ClassificationRequest):
    """
    Classify customer into segments
    """
    try:
        # TODO: Implement ML classification model
        classification = {
            "segment": "high_value",
            "confidence": 0.92,
            "characteristics": ["frequent_buyer", "high_engagement", "positive_sentiment"],
            "recommended_actions": ["premium_offers", "priority_support"]
        }
        return classification
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/v1/recommend")
async def get_recommendations(request: RecommendationRequest):
    """
    Get AI-powered recommendations for customer
    """
    try:
        # TODO: Implement recommendation engine
        recommendations = {
            "customer_id": request.customer_id,
            "recommendations": [
                {
                    "type": "product",
                    "item": "Premium Package",
                    "confidence": 0.89,
                    "reason": "Based on purchase history"
                },
                {
                    "type": "action",
                    "item": "Schedule follow-up call",
                    "confidence": 0.85,
                    "reason": "Customer engagement pattern"
                }
            ]
        }
        return recommendations
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/v1/sentiment")
async def analyze_sentiment(request: SentimentRequest):
    """
    Analyze sentiment of text
    """
    try:
        # TODO: Implement sentiment analysis model
        sentiment = {
            "text": request.text,
            "sentiment": "positive",
            "score": 0.78,
            "emotions": {
                "joy": 0.65,
                "trust": 0.72,
                "anticipation": 0.45
            },
            "key_phrases": ["excellent service", "highly recommend"]
        }
        return sentiment
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    port = int(os.getenv("AI_SERVICES_PORT", 8000))
    uvicorn.run("main:app", host="0.0.0.0", port=port, reload=True)
