ai-services/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
tensorflow==2.15.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
redis==5.0.1
httpx==0.26.0
